unit testing riset
