+++ helpers/constant.js
@@ -2,5 +2,6 @@ module.exports = {
   API_PATH: '/api/v1',
   SEARCH: {
     status: true
-  }
+  },
+  MORGAN_FORMAT: '[:date[clf]] :remote-addr :remote-user :method :url :status :res[content-length] - :response-time ms'
 }
