+++ package.json
@@ -8,6 +8,9 @@
     "dev": "nodemon --max-old-space-size=2048 --legacy-watch server.js",
     "test": "mocha",
     "coverage": "nyc npm run test",
+    "test:watch": "jest --watch",
+    "test:cov": "jest --coverage",
+    "lint": "eslint --fix --ext .js",
     "coverage-report": "nyc report --reporter=html"
   },
   "repository": {
@@ -56,7 +59,7 @@
     "method-override": "^3.0.0",
     "mocha": "^8.2.1",
     "moment": "^2.27.0",
-    "mongoose": "^5.9.25",
+    "mongoose": "^5.12.2",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.4",
     "nyc": "^15.1.0",
